[API] Further updates to the rest API test runner
