+++ elasticsearch-api/spec/elasticsearch/api/rest_api_yaml.rb
@@ -1,5 +1,24 @@
-require 'spec_helper'
+# Licensed to Elasticsearch B.V. under one or more contributor
+# license agreements. See the NOTICE file distributed with
+# this work for additional information regarding copyright
+# ownership. Elasticsearch B.V. licenses this file to you under
+# the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#	http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
 
+require 'spec_helper'
+require 'rest_yaml_tests_helper'
+
+# Match the `length` of a field.
 RSpec::Matchers.define :match_response_field_length do |expected_pairs, test|
 
   match do |response|
@@ -24,6 +43,7 @@ RSpec::Matchers.define :match_response_field_length do |expected_pairs, test|
   end
 end
 
+# Validate that a field is `true`.
 RSpec::Matchers.define :match_true_field do |field, test|
 
   match do |response|
@@ -34,6 +54,7 @@ RSpec::Matchers.define :match_true_field do |field, test|
   end
 end
 
+# Validate that a field is `false`.
 RSpec::Matchers.define :match_false_field do |field, test|
 
   match do |response|
@@ -45,6 +66,7 @@ RSpec::Matchers.define :match_false_field do |field, test|
   end
 end
 
+# Validate that a field is `gte` than a given value.
 RSpec::Matchers.define :match_gte_field do |expected_pairs, test|
 
   match do |response|
@@ -67,6 +89,7 @@ RSpec::Matchers.define :match_gte_field do |expected_pairs, test|
   end
 end
 
+# Validate that a field is `gt` than a given value.
 RSpec::Matchers.define :match_gt_field do |expected_pairs, test|
 
   match do |response|
@@ -88,6 +111,7 @@ RSpec::Matchers.define :match_gt_field do |expected_pairs, test|
   end
 end
 
+# Validate that a field is `lte` than a given value.
 RSpec::Matchers.define :match_lte_field do |expected_pairs, test|
 
   match do |response|
@@ -109,6 +133,7 @@ RSpec::Matchers.define :match_lte_field do |expected_pairs, test|
   end
 end
 
+# Validate that a field is `lt` than a given value.
 RSpec::Matchers.define :match_lt_field do |expected_pairs, test|
 
   match do |response|
@@ -130,6 +155,7 @@ RSpec::Matchers.define :match_lt_field do |expected_pairs, test|
   end
 end
 
+# Match an arbitrary field of a response to a given value.
 RSpec::Matchers.define :match_response do |expected_pairs, test|
 
   match do |response|
@@ -173,7 +199,7 @@ RSpec::Matchers.define :match_response do |expected_pairs, test|
   end
 end
 
-
+# Match that a request returned a given error.
 RSpec::Matchers.define :match_error do |expected_error|
 
   match do |actual_error|
@@ -205,6 +231,7 @@ RSpec::Matchers.define :match_error do |expected_error|
     end
   end
 end
+
 
 describe 'Rest API YAML tests' do
