+++ elasticsearch-api/spec/support/test_file/action.rb
@@ -1,3 +1,20 @@
+# Licensed to Elasticsearch B.V. under one or more contributor
+# license agreements. See the NOTICE file distributed with
+# this work for additional information regarding copyright
+# ownership. Elasticsearch B.V. licenses this file to you under
+# the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#	http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
 module Elasticsearch
 
   module RestAPIYAMLTests
@@ -34,7 +51,7 @@ module Elasticsearch
       # @param [ Elasticsearch::Client ] client The client to use to execute the action.
       # @param [ Test ] test The test containing this action. Necessary for caching variables.
       #
-      # @return [ Elasticsearch::Client ] The client. It will be a new one than the one passed in,
+      # @return [ Elasticsearch::Client ] The client. It will be a new one, not the one passed in,
       #   if the action is to set headers.
       #
       # @since 6.2.0
