+++ elasticsearch-api/spec/support/test_file/test.rb
@@ -1,3 +1,20 @@
+# Licensed to Elasticsearch B.V. under one or more contributor
+# license agreements. See the NOTICE file distributed with
+# this work for additional information regarding copyright
+# ownership. Elasticsearch B.V. licenses this file to you under
+# the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#	http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
 module Elasticsearch
 
   module RestAPIYAMLTests
@@ -170,14 +187,9 @@ module Elasticsearch
 
           if @skip
             @skip.collect { |s| s['skip'] }.any? do |skip|
-              contains_features_to_skip?(features_to_skip, skip) ||
-                  skip_version?(client, skip)
+              contains_features_to_skip?(features_to_skip, skip) || skip_version?(client, skip)
             end
           end
-        end
-
-        def convert_response(response)
-          binding.pry
         end
 
         private
